+++ Source/Lib/Common/Codec/EbMotionEstimationProcess.c
@@ -379,8 +379,9 @@ EbErrorType motion_estimation_context_ctor(
 /***************************************************************************************************
 * ZZ Decimated SAD Computation
 ***************************************************************************************************/
+#if !ADAPTIVE_QP_SCALING
 int non_moving_th_shift[4] = { 4, 2, 0, 0 };
-
+#endif
 EbErrorType ComputeDecimatedZzSad(
     MotionEstimationContext_t   *context_ptr,
     SequenceControlSet        *sequence_control_set_ptr,
@@ -493,8 +494,17 @@ EbErrorType ComputeDecimatedZzSad(
 #endif
                 decimatedLcuCollocatedSad = (uint32_t)~0;
             }
-
-
+#if ADAPTIVE_QP_SCALING
+            // Keep track of non moving LCUs for QP modulation
+            if (decimatedLcuCollocatedSad < ((decimatedLcuWidth * decimatedLcuHeight) * 2)) 
+                previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_0_ZZ_COST;
+            else if (decimatedLcuCollocatedSad < ((decimatedLcuWidth * decimatedLcuHeight) * 4)) 
+                previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_1_ZZ_COST;
+            else if (decimatedLcuCollocatedSad < ((decimatedLcuWidth * decimatedLcuHeight) * 8)) 
+                previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_2_ZZ_COST;
+            else 
+                previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_3_ZZ_COST;
+#else
             // Keep track of non moving LCUs for QP modulation
             if (decimatedLcuCollocatedSad < ((decimatedLcuWidth * decimatedLcuHeight) * 2) >> non_moving_th_shift[sequence_control_set_ptr->input_resolution]) {
                 previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_0_ZZ_COST;
@@ -508,6 +518,7 @@ EbErrorType ComputeDecimatedZzSad(
             else { 
                 previous_picture_control_set_wrapper_ptr->non_moving_index_array[sb_index] = BEA_CLASS_3_ZZ_COST;
             }
+#endif
         }
     }