+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Static library support  
 - Alt-ref pictures - temporal filtering
 - Adaptive Transform Block for INTRA
+- Adaptive QP scaling
 
 ## [0.5.0] - 2019-05-17